+++ src/main/java/buttondevteam/lib/TBMCCoreAPI.java
@@ -18,6 +18,7 @@ import org.bukkit.plugin.Plugin;
 import com.google.gson.*;
 
 import buttondevteam.core.MainPlugin;
+import buttondevteam.lib.potato.DebugPotato;
 
 public final class TBMCCoreAPI {
 	static List<String> coders = new ArrayList<String>() {
