+++ src/main/java/buttondevteam/lib/player/ChromaGamerBase.java
@@ -1,4 +1,4 @@
-package buttondevteam.lib;
+package buttondevteam.lib.player;
 
 import org.bukkit.configuration.ConfigurationSection;
