+++ src/main/java/buttondevteam/lib/player/TBMCPlayerBase.java
@@ -1,4 +1,4 @@
-package buttondevteam.lib;
+package buttondevteam.lib.player;
 
 import java.util.UUID;
