+++ src/main/java/buttondevteam/lib/player/TBMCPlayerSaveEvent.java
@@ -1,4 +1,4 @@
-package buttondevteam.lib;
+package buttondevteam.lib.player;
 
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.event.Event;
