+++ src/main/java/buttondevteam/lib/potato/Gary.java
@@ -1,4 +1,4 @@
-package buttondevteam.lib;
+package buttondevteam.lib.potato;
 
 import java.util.List;
